+++ tests/testthat/test-cfbd_calendar.R
@@ -7,7 +7,7 @@ cols <- c(
 test_that("CFB Calendar", {
   skip_on_cran()
   x <- cfbd_calendar(year = 2019)
-  expect_equal(nrow(x), 16)
+  expect_equal(nrow(x), 17)
   expect_equal(colnames(x), cols)
   expect_s3_class(x, "data.frame")
 })