+++ NEWS.md
@@ -1,3 +1,9 @@
+# **cfbfastR v1.9.1**
+
+* Improved drive_pts logic in play-by-play data.
+* Fixed an issue that occasionally made the `cfbd_game_team_stats()` function return data in a long format
+* Minor documentation and test updates
+
 # **cfbfastR v1.9.0**
 
 #### Added functions to access ESPN API:
@@ -14,6 +20,7 @@
 * [```load_cfb_teams()```](https://cfbfastR.sportsdataverse.org/reference/load_cfb_teams.html)
 
 - Removes `furrr`, `future` dependencies, adds `Rcpp`, `RcppParallel`, and `purrr` dependencies
+
 
 # **cfbfastR v1.8.0**